requirements.txt
python-dotenv
requests
pandas
numpy
pydantic

openai
langchain
langgraph

faiss-cpu
sentence-transformers

yfinance
newsapi-python

aiohttp
rich

google-adk
google-generativeai